split tracking service
