+++ apps/web/pages/api/claim.ts
@@ -5,7 +5,7 @@ import { validateRequest } from '../../services/security';
 import { getBlockchainService } from '../../services/blockchains';
 import FaucetService from '../../services/faucet/faucetService';
 import { Network, getAppConfig } from '@config';
-import { getTransactionHistoryService } from '../../services/transactionHistory';
+import { getTrackingService } from '@trackingService';
 import { getErrorMessage } from '@/utils';
 
 export type ClaimParams = {
@@ -21,8 +21,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse<DefaultResponse
   const blockchainService = getBlockchainService(networkName);
   try {
     await validateRequest(req);
-    const { transactionHistoryType } = getAppConfig();
-    const transactionHistoryService = getTransactionHistoryService(transactionHistoryType);
+    const { trackingType } = getAppConfig();
+    const transactionHistoryService = getTrackingService(trackingType);
     const faucetService = new FaucetService(blockchainService, transactionHistoryService);
 
     const { address, message, signature }: ClaimParams = req.body;
