+++ apps/web/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './ethAddressUtils';
 export * from './textMessage';
-export * from './transactionHistory';
