+++ libs/tracking-service/README.md
@@ -1,0 +1,7 @@
+# tracking-service
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build tracking-service` to build the library.
