+++ libs/tracking-service/src/interfaces.ts
@@ -1,5 +1,5 @@
 import { Network } from '@config';
-export interface ITransactionHistoryService {
+export interface ITrackingService {
   hasReceivedTokens(network: Network, address: string, minLayover?: number): Promise<boolean>;
   recordTransaction(network: Network, address: string): Promise<void>;
 }
