+++ tsconfig.base.json
@@ -18,7 +18,8 @@
       "@/*": ["./apps/web/*"],
       "@config": ["libs/config/src/index.ts"],
       "@errors": ["libs/errors/src/index.ts"],
-      "@nonceService": ["libs/nonce-service/src/index.ts"]
+      "@nonceService": ["libs/nonce-service/src/index.ts"],
+      "@trackingService": ["libs/tracking-service/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
